refactor: move monitoring functions to separate module
